refactor: :zap: refactor routes
