+++ test_e2e/customers/get/200_get_by_id.test.ts
@@ -1,5 +1,5 @@
 import request from "supertest";
-import { app, pool } from "../../app";
+import { app, pool } from "../../../src/app";
 
 afterAll(() => {
 	pool.end();
